CyberBodyguard/requirements.txt
customtkinter
pymongo
requests
watchdog
python-dotenv
plyer
Pillow
pyinstaller